Move source to core module.
