+++ core/src/main/java/de/felix_klauke/caroline/core/CarolineApplication.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package de.felix_klauke.caroline;
+package de.felix_klauke.caroline.core;
 
 import io.reactivex.Scheduler;
 import io.reactivex.plugins.RxJavaPlugins;
