+++ core/src/main/java/de/felix_klauke/caroline/core/scheduler/AbstractScheduler.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package de.felix_klauke.caroline.scheduler;
+package de.felix_klauke.caroline.core.scheduler;
 
 import io.reactivex.Scheduler;
 import io.reactivex.disposables.CompositeDisposable;
